Fix bug where checking over 2048M on 64-bit systems failed. Updated changelog
git-svn-id: https://svn.upfronthosting.co.za/svn/upfront/MonitorTools/trunk@102 80044c55-6f29-0410-b7d5-9c2a68987f97

+++ nagios/check_process_memory.c
@@ -65,10 +65,10 @@ int main(int argc, char **argv){
                 target = optarg;
                 break;
             case 'W':
-                warning = atol(optarg)*(strchr(optarg, 'k')?1024:(strchr(optarg, 'm')?1048576:1));
+                warning = (unsigned long)atol(optarg)*(strchr(optarg, 'k')?1024:(strchr(optarg, 'm')?1048576:1));
                 break;
             case 'C':
-                critical = atol(optarg)*(strchr(optarg, 'k')?1024:(strchr(optarg, 'm')?1048576:1));
+                critical = (unsigned long)atol(optarg)*(strchr(optarg, 'k')?1024:(strchr(optarg, 'm')?1048576:1));
                 break;
             case 'P':
                 cmdline = optarg;
